IN-196 refactor: fix imports, model and configuration
