+++ src/common/api/index.ts
@@ -1,0 +1,1 @@
+export * from './locations.api'
