+++ src/common/api/locations.api.ts
