+++ src/common/index.ts
@@ -1,3 +1,4 @@
+export * from './api'
 export * from './components'
 export * from './enviroments'
 export * from './form'
