+++ src/features/createPost/index.ts
@@ -1,3 +1,2 @@
-// export * from './api'
 export * from './model'
 export * from './ui'
