+++ src/features/createPost/model/adding/index.ts
@@ -1,2 +1,3 @@
 export * from './addPhotoSchema'
 export * from './addPhotosCheck'
+export * from './useAddPhotoModal'
