+++ src/features/createPost/model/index.ts
@@ -1,6 +1,5 @@
 export * from './addPostStore'
 export * from './adding'
-export * from './constants'
 export * from './cropping'
 export * from './publication'
 export * from './types'
