+++ src/features/createPost/model/types.ts
@@ -2,10 +2,6 @@ import { ElementType } from 'react'
 import { Area, Point } from 'react-easy-crop'
 
 import { Nullable, PhotoResult } from '@/common'
-import { PhotoEditorState } from '@/features/createPost/model/constants'
-
-export type PhotoEditorStateType =
-  (typeof PhotoEditorState)[keyof typeof PhotoEditorState]
 
 export type PostPhoto = {
   aspect: number
