+++ src/features/profile/settings/generalInfo/api/index.ts
@@ -1,3 +1,3 @@
-export * from './locations.api'
 export * from './profile.api'
 export * from './profileApi.types'
+export * from 'src/common/api/locations.api'
